requirements.txt
eyeD3==0.8.10
get==2019.4.13
post==2019.4.13
public==2019.4.13
python-magic==0.4.15
query-string==2019.4.13
request==2019.4.13
six==1.12.0
youtube-dl==2019.9.1
